Fixed quickSort issue and changed index.js timeValue function to a case switch function
Found out today that my quicksort function was not properly quicksorting the data and would change everytime the function ran. This was cause by not putting in the sortBy parameter into the recursion functions at the end leading to the next time it fired itself it would get an array or undefined when working out the pivot and values of each array to compare to one another.

I also changed the timeValue function in index.js which works out a aprox value of time from the postedAt string that puppeteer scrapes. I changed this to a switch case function to increase readability and to reduce the amount of if else statements.